# server/api/swagger/swagger.yaml
swagger: "2.0"
info:
  description: "A assistant for wanfang team"
  version: "1.0.0"
  title: "WFTeam"
  contact:
    email: "hueidou163@gmail.com"
  license:
    name: "MIT License"
    url: "https://github.com/hueidou/WFTeam/blob/master/LICENSE"
host: "localhost:8000"
basePath: "/api"
tags:
- name: "user"
  description: "用户"
schemes:
- "http"
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /hello:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: hello
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
  /user:
    get:
      tags:
      - "user"
      x-swagger-router-controller: "User"
      summary: "获取用户列表"
      description: ""
      operationId: "getUsers"
      produces:
      - "application/json"
      responses:
        200:
          description: ""
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
  /user/login:
    post:
      tags:
      - "user"
      x-swagger-router-controller: "User"
      summary: "登录"
      description: ""
      operationId: "loginUser"
      produces:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        description: "The user name for login"
        required: true
        schema:
          $ref: "#/definitions/LoginUser"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/LoginUser"
          headers:
            X-Rate-Limit:
              type: "integer"
              format: "int32"
              description: "calls per hour allowed by the user"
            X-Expires-After:
              type: "string"
              format: "date-time"
              description: "date in UTC when token expires"
        400:
          description: "Invalid username/password supplied"
  /user/profile:
    put:
      tags:
      - "user"
      x-swagger-router-controller: "User"
      summary: "完善用户信息"
      description: ""
      operationId: "profileUser"
      produces:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        description: "The user name for login"
        required: true
        schema:
          $ref: "#/definitions/ProfileUser"
      responses:
        200:
          description: "OK"
  /project:
    get:
      tags:
      - "project"
      x-swagger-router-controller: "Project"
      summary: "获取项目列表"
      description: ""
      operationId: "getProjects"
      produces:
      - "application/json"
      responses:
        200:
          description: "OK"
          schema:
            type: array
            items:
              $ref: '#/definitions/Project'
    post:
      tags:
      - "project"
      x-swagger-router-controller: "Project"
      summary: "新增项目"
      description: ""
      operationId: "addProject"
      produces:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/Project"
      responses:
        200:
          description: "新增项目成功"
        400:
          description: "参数不合法"
  /manday:
    get:
      tags:
      - "manday"
      x-swagger-router-controller: "Manday"
      summary: "获取工时列表"
      description: ""
      operationId: "getMandays"
      produces:
      - "application/json"
      parameters:
      - name: "show"
        in: "query"
        description: ""
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: array
            items:
              $ref: '#/definitions/Manday'
    post:
      tags:
      - "manday"
      x-swagger-router-controller: "Manday"
      summary: "新增工时"
      description: ""
      operationId: "addManday"
      produces:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/Manday"
      responses:
        200:
          description: "新增工时成功"
        400:
          description: "参数不合法"
  /manday/{_id}:
    delete:
      tags:
      - "manday"
      x-swagger-router-controller: "Manday"
      summary: "删除工时"
      description: ""
      operationId: "deleteManday"
      produces:
      - "application/json"
      parameters:
      - name: "_id"
        in: "path"
        description: ""
        required: true
        type: "string"
      responses:
        200:
          description: "删除工时成功"
    put:
      tags:
      - "manday"
      x-swagger-router-controller: "Manday"
      summary: "编辑工时"
      description: ""
      operationId: "modifyManday"
      produces:
      - "application/json"
      parameters:
      - name: "_id"
        in: "path"
        description: ""
        required: true
        type: "string"
      - name: "body"
        in: "body"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/Manday"
      responses:
        200:
          description: "更新工时成功"
        400:
          description: "参数不合法"
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  LoginUser:
    type: "object"
    properties:
      userName:
        type: "string"
  ProfileUser:
    type: "object"
    properties:
      realName:
        type: "string"        
  User:
    type: "object"
    properties:
      userName:
        type: "string"
      realName:
        type: "string"
  Project:
    type: "object"
    properties:
      projectKey:
        type: "string"
      projectName:
        type: "string"
  Manday:
    type: "object"
    properties:
      date:
        type: "string"
        format: "date-time"
      projectKey:
        type: "string"
      hours:
        type: "integer"
        format: "int32"
      description:
        type: "string"
externalDocs:
  description: "接口说明"
  url: "https://github.com/hueidou/WFTeam/blob/master/server/README.md"